+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix missing chevron image in quickaccess
 - Remove EJS from dependencies
 - Fix import of keepass file containing entries with undefined field
+- Fix import should not throw an error if a resource or a folder cannot be created
 
 ## [3.0.5] - 2021-02-03
 ### Fixed